requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
polib==1.2.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==25.1.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.17.0
stack_data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
zipp==3.21.0
